Move test to follow same folder structure as code
