+++ README.md
@@ -118,7 +118,7 @@ Each module declares its own Alembic branch label in its initial migration. This
 For example, the members module uses branch `members_core`:
 
 ```
-src/modules/members/migrations/versions/
+src/cofy/modules/members/migrations/versions/
 ├── members_core_0001_members_core_initial.py   # branch_labels = ("members_core",)
 └── members_core_0002_add_phone_number.py       # extends members_core@head
 ```
@@ -126,7 +126,7 @@ src/modules/members/migrations/versions/
 A separate module `foo` would have its own branch `foo_core` with revisions in its own directory:
 
 ```
-src/modules/foo/migrations/versions/
+src/cofy/modules/foo/migrations/versions/
 ├── foo_core_0001_initial.py                    # branch_labels = ("foo_core",)
 └── foo_core_0002_add_index.py                  # extends foo_core@head
 ```
