+++ pyproject.toml
@@ -24,6 +24,9 @@ build-backend = "setuptools.build_meta"
 [tool.setuptools.packages.find]
 where = ["src"]
 namespaces = true
+
+[tool.setuptools.package-data]
+"cofy.db" = ["alembic/*", "alembic/**/*"]
 [dependency-groups]
 dev = [
     "pytest-asyncio>=1.3.0",
