+++ tests/cofy/api/docs_router_test.py
