+++ tests/cofy/modules/tariff/sources/entsoe_day_ahead_example.csv
