+++ tests/cofy/modules/tariff/sources/entsoe_day_ahead_test.py
@@ -11,7 +11,7 @@ from cofy.modules.tariff import EntsoeDayAheadTariffSource
 from cofy.modules.timeseries import Timeseries
 
 EXAMPLE_CSV_NAME = "entsoe_day_ahead_example.csv"
-EXAMPLE_CSV_PATH = resources.files("tests.modules.tariff.sources").joinpath(EXAMPLE_CSV_NAME)
+EXAMPLE_CSV_PATH = resources.files("tests.cofy.modules.tariff.sources").joinpath(EXAMPLE_CSV_NAME)
 
 
 def test_init_valid():
