+++ tests/cofy/modules/timeseries/formats/csv_test.py
